fix: Deletion resources related to subgroup (#95)
Problem:
If we delete the subgroup before deleting the resources
that contain this subgroup
(Example: KeycloakRealmGroup.spec.subGroup/KeycloakRealmUser.spec.groups)
deletion of KeycloakRealmGroup/KeycloakRealmUser blocks with an error.
This happens because we are syncing KeycloakRealmUser/KeycloakRealmGroup
before deletion, and syncing can't be processed without a subgroup.
Solution:
Moved deletion of KeycloakRealmUser/KeycloakRealmGroup before syncing.
We don't need to sync objects that are marked for deletion.